update: code compiles

+++ shared_types/src/operational/operational_response_status.rs
@@ -5,19 +5,19 @@ use crate::scheduling_environment::worker_environment::resources::Id;
 #[derive(Serialize)]
 pub struct OperationalStatusResponse {
     id: Id,
-    assign_number_of_activities: usize,
-    assess_number_of_activities: usize,
-    unassign_number_of_activities: usize,
-    objective: usize,
+    assign_number_of_activities: u64,
+    assess_number_of_activities: u64,
+    unassign_number_of_activities: u64,
+    objective: u64,
 }
 
 impl OperationalStatusResponse {
     pub fn new(
         id: Id,
-        assign_number_of_activities: usize,
-        assess_number_of_activities: usize,
-        unassign_number_of_activities: usize,
-        objective: usize,
+        assign_number_of_activities: u64,
+        assess_number_of_activities: u64,
+        unassign_number_of_activities: u64,
+        objective: u64,
     ) -> Self {
         Self {
             id,
